+++ officebot.js
@@ -1,7 +1,8 @@
-const config = require("./src/config.json");
+// Auto functions
 const autoActivity = require("./src/js/auto-functions/auto-activity");
 const autoReadCommand = require("./src/js/auto-functions/auto-read-command");
-
+// Config
+const config = require("./src/config.json");
 const discord = require("discord.js");
 const bot = new discord.Client();