+++ src/js/items/commands.js
@@ -4,6 +4,8 @@ module.exports = [
     {commands: ['createvoicechannel', 'cvc'], description: 'Create voice channel', arguments: '\<\channel-name>+<category-id> \n required+n-required'},
     {commands: ['createtextchannel', 'ctc'], description: 'Create text channel', arguments: '\<\channel-name>+<category-id> \n required+n-required'},
     {commands: ['clearchannel', 'cc'], description: 'Clear channel'},
+    {commands: ['kick'], description: 'Kick member', arguments: '\<\@member> \n required'},
+    {commands: ['ban'], description: 'Ban member', arguments: '\<\@member> \n required'},
     {commands: ['ping'], description: 'Pong!'},
 ]