+++ src/js/main-functions/create-help.js
@@ -1,7 +1,11 @@
+// Nano functions
 const createMessageEmbed = require("../nano-functions/create-message-embed");
+const addReactions = require("../nano-functions/add-reactions");
+// Config
 const { PREFIX } = require("../../config.json");
+const _ = require("lodash");
 
-module.exports = (bot, message, commands, _) => {
+module.exports = (bot, message, commands) => {
     let fields = _.cloneDeep(commands);
 
     fields.forEach((field) => {
@@ -13,5 +17,15 @@ module.exports = (bot, message, commands, _) => {
     });
 
     message.channel.send(createMessageEmbed(`Commands list (${PREFIX}command)`, message.guild.iconURL(), fields));
-    message.delete();
+    const react = async (message) => {
+        await message.delete();
+
+        const fetched = await message.channel.messages.fetch({ limit: 1 });
+
+        if (fetched && fetched.first()) {
+            addReactions(fetched.first(), ['⬅', '➡']);
+        }
+    };
+
+    react(message);
 }