+++ src/js/nano-functions/add-reactions.js
@@ -1,7 +1,6 @@
 module.exports = (message, reactions) => {
-    message.react(reactions[0]);
-    reactions.shift();
-    if (reactions.length > 0) {
-        setTimeout(() => addReactions(message, reactions), 750)
+    while (reactions.length > 0) {
+        message.react(reactions[0]);
+        reactions.shift();
     }
 };