+++ src/js/nano-functions/create-message-embed.js
@@ -1,3 +1,4 @@
+// Config
 const { MessageEmbed } = require("discord.js");
 
 module.exports = (title, icon, fields) => {