+++ src/js/nano-functions/split-args.js
@@ -1,3 +1,4 @@
+// Config
 const { SEPARATOR_ARGS } = require("../../config.json");
 
 module.exports = (args) => {